# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MotivateRequest {
  steps: number;
  durationSec: number;
  label: string;
}

export interface MotivateResponse {
  text: string;
}

export interface InsightRequest {
  successCount: number;
  totalCount: number;
  avgSeconds: number;
  bestSeconds: number;
}

export interface InsightResponse {
  text: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
  details?: string[];
}